moved Player class to new module player.py
